Adjust search path.
When following the instructions in the README file, the ghci complains
about a mismatch between file name and module name. It expects each
module to start with 'HAD' because the search path is set to 'exercises'
instead of 'excercises/HAD'. This commit fixes the issue by adjusting
the search path.